[mod] #17 파일명 수정 및 약서랍 activity -> fragment로 수정
